# server-codegeneration/openapi.yaml
openapi: 3.0.1
info:
  title: Hotel Service API
  version: 1.0.1
  description: API for managing guests, rooms, and reservations in a hotel.

servers:
  - url: http://localhost:8080
    description: Local server

paths:
  /guests:
    get:
      tags:
        - guest
      summary: Get all guests
      operationId: getGuests
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
          description: The name of the guest to search for
      responses:
        '200':
          description: A list of guests, optionally filtered by name
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Guest'
    post:
      summary: Create a new guest
      tags:
        - guest
      operationId: createGuest
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuestCreate'
      responses:
        '201':
          $ref: '#/components/responses/CreatedWithLocation'

  /guests/{guestId}:
    get:
      summary: Get a guest by ID
      tags:
        - guest
      operationId: getGuestById
      parameters:
        - name: guestId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: The guest with the specified ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Guest'
        '404':
          description: Guest not found
    patch:
      summary: Partially update a guest by ID
      tags:
        - guest
      operationId: updateGuestById
      parameters:
        - name: guestId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The ID of the guest to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuestUpdate'
      responses:
        '200':
          description: Guest updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Guest'
        '404':
          description: Guest not found
        '400':
          description: Invalid input

    delete:
      summary: Delete a guest by ID
      tags:
        - guest
      operationId: deleteGuestById
      parameters:
        - name: guestId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Guest deleted successfully
        '404':
          description: Guest not found

  /guests/{guestId}/reservations:
    get:
      summary: Get all reservations for a specific guest
      tags:
        - reservation
      operationId: getReservationsByGuest
      parameters:
        - name: guestId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The ID of the guest to retrieve reservations for
      responses:
        '200':
          description: A list of reservations for the specified guest
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reservation'
        '404':
          description: Guest not found or no reservations for this guest

  /rooms:
    get:
      summary: Get all rooms
      tags:
        - room
      operationId: getRooms
      responses:
        '200':
          description: A list of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Room'
  /rooms/{roomNumber}:
    get:
      summary: Get a room by room number
      tags:
        - room
      operationId: getRoomByNumber
      parameters:
        - name: roomNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The room with the specified number
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '404':
          description: Room not found
  /rooms/{roomNumber}/reservations:
    get:
      summary: Get all reservations for a specific room
      tags:
        - reservation
      operationId: getReservationsByRoom
      parameters:
        - name: roomNumber
          in: path
          required: true
          schema:
            type: string
          description: The number of the room to retrieve reservations for
      responses:
        '200':
          description: A list of reservations for the specified room
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reservation'
        '404':
          description: Room not found or no reservations for this room

  /reservations:
    get:
      summary: Get all reservations
      tags:
        - reservation
      operationId: getReservations
      responses:
        '200':
          description: A list of reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reservation'
    post:
      summary: Create a new reservation
      tags:
        - reservation
      operationId: createReservation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservationCreate'
      responses:
        '201':
          $ref: '#/components/responses/CreatedWithLocation'

  /reservations/{reservationId}:
    get:
      summary: Get a reservation by ID
      tags:
        - reservation
      operationId: getReservationById
      parameters:
        - name: reservationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: The reservation with the specified ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        '404':
          description: Reservation not found
    put:
      summary: Update a reservation by ID
      tags:
        - reservation
      operationId: updateReservationById
      parameters:
        - name: reservationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Reservation'
      responses:
        '200':
          description: Reservation updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        '404':
          description: Reservation not found
    delete:
      summary: Delete a reservation by ID
      tags:
        - reservation
      operationId: deleteReservationById
      parameters:
        - name: reservationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Reservation deleted successfully
        '404':
          description: Reservation not found

  /_search:
    post:
      summary: Search across guests, rooms, and reservations
      tags:
        - search
      operationId: searchAll
      requestBody:
        description: Search criteria for guests, rooms, and reservations
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchCriteria'
      responses:
        '200':
          description: Search results across multiple resources
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResults'

components:
  responses:
    CreatedWithLocation:
      description: Resource created successfully
      headers:
        Location:
          description: URI of the newly created resource
          schema:
            type: string
  schemas:
    GuestBase:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
      required:
        - name
        - email

    GuestCreate:
      allOf:
        - $ref: '#/components/schemas/GuestBase'

    Guest:
      allOf:
        - $ref: '#/components/schemas/GuestBase'
        - type: object
          properties:
            id:
              type: integer
              format: int64
          required:
            - id
            - createdAt

    GuestUpdate:
      type: object
      properties:
        name:
          type: string
          description: The new name of the guest
        email:
          type: string
          format: email
          description: The new email of the guest
        phoneNumber:
          type: string
          description: The new phone number of the guest

    Room:
      type: object
      properties:
        number:
          type: string
        type:
          type: string
          enum: [ Single, Double, Suite ]
        capacity:
          type: integer
        status:
          type: string
          enum: [ Available, Booked ]
        pricePerNight:
          type: number
          format: float
      required:
        - number
        - type

    ReservationBase:
      type: object
      properties:
        guestId:
          type: integer
          format: int64
        roomNumber:
          type: string
        checkIn:
          type: string
          format: date
        checkOut:
          type: string
          format: date
      required:
        - guestId
        - roomNumber
        - checkIn
        - checkOut

    ReservationCreate:
      allOf:
        - $ref: '#/components/schemas/ReservationBase'

    Reservation:
      allOf:
        - $ref: '#/components/schemas/ReservationBase'
        - type: object
          properties:
            id:
              type: integer
              format: int64
          required:
            - id

    SearchCriteria:
      type: object
      properties:
        guestName:
          type: string
          description: Search by guest's name
        roomType:
          type: string
          description: Search by room type (Single, Double, Suite)
        reservationDate:
          type: string
          format: date
          description: Search reservations by check-in or check-out date

    SearchResults:
      type: object
      properties:
        guests:
          type: array
          items:
            $ref: '#/components/schemas/Guest'
        rooms:
          type: array
          items:
            $ref: '#/components/schemas/Room'
        reservations:
          type: array
          items:
            $ref: '#/components/schemas/Reservation'